made RoR and contributions proper child components of Performance
